+++ src/Remap.cc
@@ -117,7 +117,7 @@ int maptype2operfunc(int map_type, int submap_type, int num_neighbors, int remap
 } 
 
 static
-void print_remap_info(int operfunc, int remap_genweights, remapgrid_t *src_grid, remapgrid_t *tgt_grid, int nmiss)
+void print_remap_info(int operfunc, int remap_genweights, remapgrid_t *src_grid, remapgrid_t *tgt_grid, size_t nmiss)
 {
   char line[256], tmpstr[256];
 
@@ -163,7 +163,7 @@ void print_remap_info(int operfunc, int remap_genweights, remapgrid_t *src_grid,
 }
 
 static
-void print_remap_warning(const char *remap_file, int operfunc, remapgrid_t *src_grid, int nmiss)
+void print_remap_warning(const char *remap_file, int operfunc, remapgrid_t *src_grid, size_t nmiss)
 {
   char line[256];
   char tmpstr[256];
@@ -780,7 +780,7 @@ void *Remap(void *argument)
   int varID, levelID;
   size_t gridsize, gridsize2;
   int gridID1 = -1, gridID2;
-  int nmiss1, nmiss2;
+  size_t nmiss1, nmiss2;
   size_t i, j;
   int r = -1;
   int nremaps = 0;
