+++ src/fieldmer.cc
@@ -326,7 +326,7 @@ void meravgw(field_type field1, field_type *field2)
 }
 
 static
-void prevarsum_merw(const double *restrict array, const double *restrict w, int nx, int ny, int nmiss, 
+void prevarsum_merw(const double *restrict array, const double *restrict w, int nx, int ny, size_t nmiss, 
                     double missval, double *restrict rsum, double *restrict rsumw, double *restrict rsumq, double *restrict rsumwq)
 { 
   *rsum   = 0;
