+++ src/fieldzon.cc
@@ -322,7 +322,7 @@ void zonavg(field_type field1, field_type *field2)
 }
 
 static
-void prevarsum_zon(const double *restrict array, int nx, int nmiss,  double missval, 
+void prevarsum_zon(const double *restrict array, int nx, size_t nmiss,  double missval, 
                    double *rsum, double *rsumw, double *rsumq, double *rsumwq)
 {
   double w = 1./nx;
